% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPPairedSamples.R
\name{contours.plot}
\alias{contours.plot}
\title{Contour Plot Function}
\usage{
contours.plot(results_BNP)
}
\arguments{
\item{results_BNP}{list of results obtained from the \code{BNP.test} function.}
}
\value{
It returns an interactive graph showing the contours and the points representing
each pair of sample data. By scrolling within the graph it is possible to observe
 in detail the coordinates that delimit the contour and the density they represent.
}
\description{
The contour plot allows to visualize the joint distribution of the data,
as well as to identify the correlation, if any, between the two study populations.
}
\examples{

\dontrun{
results <- BNP.test(x1, y1, n.mcm=10000)
contours.plot(results)
}

}
