% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPPairedSamples.R
\name{plotshift.function}
\alias{plotshift.function}
\title{Shift Plot Function}
\usage{
plotshift.function(results_BNP)
}
\arguments{
\item{results_BNP}{list of results obtained from the \code{BNP.test} function.}
}
\value{
It returns an interactive graph with the shift function estimate and the
credible interval at 95% confidence. By scrolling within the graph it is
possible to observe in detail the grid of 200 values ranging from the
minimum of the data to the maximum; it also shows the average of the
differences between the quantiles of both distributions.
}
\description{
The shift function quantifies the differences between the two study
populations, before and after the intervention or measurement, through
the difference between the quantiles of the two distributions as a
function of the quantiles of the first group.
}
\note{
It is suggested to use the shift function if significant
differences are observed between the marginal distributions
(\code{posterior.probability.H1}), after performing the
hypothesis test \code{BNP.test}.
}
\examples{

results <- BNP.test(x1, y1, n.mcm=10000)
plot.shift.function(results)


}
